requirements.txt
robotframework>=7.4
pydantic>=2.0
openai>=2.15.0
robotframework-browser>=19.12.3